Extract client-side code to separate components
