+++ app/auth/sent/RetryButton.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { signIn } from "next-auth/react";
+
+export const RetryButton = ({ callbackUrl }: { callbackUrl: string }) => (
+  <a
+    className="btn-inverted m-auto inline-block"
+    onClick={() => signIn(undefined, { callbackUrl })}
+  >
+    Zkusit znova
+  </a>
+);
