+++ app/auth/sent/page.tsx
@@ -1,8 +1,6 @@
-"use client";
-
 import { Fragment } from "react";
 
-import { signIn } from "next-auth/react";
+import { RetryButton } from "./RetryButton";
 
 type Props = {
   searchParams: {
@@ -24,12 +22,7 @@ const Page = ({ searchParams }: Props) => {
         </a>
         , pomůžeme.
       </p>
-      <a
-        className="btn-inverted m-auto inline-block"
-        onClick={() => signIn(undefined, { callbackUrl })}
-      >
-        Zkusit znova
-      </a>
+      <RetryButton callbackUrl={callbackUrl} />
     </Fragment>
   );
 };
